rag_single_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "text2sql"
version = "0.1.0"
description = "Text-to-SQL Agent Platform for Banking/POS"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "asyncpg>=0.29",
    "chromadb>=0.5",
    "sentence-transformers>=2.2",
    "sqlparse>=0.5",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "streamlit>=1.38",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests requiring real Claude API and PostgreSQL (deselect with '-m not e2e')",
]
addopts = "-m 'not e2e'"

[tool.ruff]
target-version = "py311"
line-length = 120
